Docu clean-up with @brief

+++ SignalMonitor.cpp
@@ -1,14 +1,19 @@
 #include "SignalMonitor.h"
 
-
-SignalMonitor::SignalMonitor(int _x_kn1)
+/**
+ * Construct the signal monitor object
+ */
+SignalMonitor::SignalMonitor(float _x_kn1)
 {
     x_kn1 = _x_kn1;
 }
 
-
-/* returns 1 if signal increased */
-boolean SignalMonitor::detectIncrease(int _x_k)
+/**
+ * @brief Detect increase
+ * 
+ * @return 1 if input signal has increased 
+ */
+boolean SignalMonitor::detectIncrease(float _x_k)
 {
     boolean result = 0;
     if(_x_k > x_kn1) result = 1;
@@ -17,8 +22,12 @@ boolean SignalMonitor::detectIncrease(int _x_k)
     return(result);
 }
 
-/* returns 1 if signal decreased */
-boolean SignalMonitor::detectDecrease(int _x_k)
+/**
+ * @brief Detect decrease
+ * 
+ * @return 1 if input signal has decreased 
+ */
+boolean SignalMonitor::detectDecrease(float _x_k)
 {
     boolean result = 0;
     if(_x_k < x_kn1) result = 1;
@@ -27,8 +36,12 @@ boolean SignalMonitor::detectDecrease(int _x_k)
     return(result);
 }
 
-/* returns 1 if signal has changed */
-boolean SignalMonitor::detectChange(int _x_k)
+/**
+ * @brief Detect change
+ * 
+ * @return 1 if input signal has changed 
+ */
+boolean SignalMonitor::detectChange(float _x_k)
 {
     boolean result = 0;
     if(_x_k != x_kn1) result = 1;
@@ -36,11 +49,3 @@ boolean SignalMonitor::detectChange(int _x_k)
     x_kn1 = _x_k;
     return(result);
 }
-
-
-
-
-
-
-
-
