+++ SignalMonitor.h
@@ -3,7 +3,10 @@
 
 #include <Arduino.h>
 
-/* The SignalMonitor is watching an integer software-signal and detects changes
+/** 
+ * @brief Monitoring of a software signal to detect increase, decrease or change. 
+ * 
+ * The SignalMonitor is watching an integer software-signal and detects changes
  * 
  * @author E. M. Luebeck
  * @date 2021-04-23
@@ -11,13 +14,13 @@
 class SignalMonitor 
 {
     public:
-        SignalMonitor(int _x_kn1);
-        boolean detectIncrease(int _x_k);
-        boolean detectDecrease(int _x_k);
-        boolean detectChange(int _x_k);
+        SignalMonitor(float _x_kn1);
+        boolean detectIncrease(float _x_k);
+        boolean detectDecrease(float _x_k);
+        boolean detectChange(float _x_k);
         
     private:
-        int x_kn1;
+        float x_kn1;
 };
 
 #endif /* SIGNALMONITOR_H */
